Support for limiting editing of attributes
Content controllers is a new feature for controlling which
attributes can be edited by the current user.
Each content controller is called when fetching input for an
attribute and may deny access to editing that attribute.

A content controller is defined by adding the class name to
EditSettings/ContentControllers in content.ini. If no controllers
are defined then all attributes are allowed as before.
The class must then implement the `canEditAttribute` method
which accepts the object attribute as the first parameter.