Changed all access to var to go via eZSys methods
This ensures that var and var/cache can be overriden with
environment variables, for instance to move all cache to
a common folder.

Added $trailing parameter to all path/dir methods in eZSys,
to either add a trailing slash or remove it.

Added new path/dir methods to eZSys, autoloadPath() and
externalAutoloadPath().